[Feature/#185] 배포환경에서 알림 안오는 이슈해결 (#191)
* fix: 알림 안오는 문제- console.log 추가, 카카오공유 정보추가

* fix: 링크 공유 url 수정

* fix: firebase config 파일 빌드시 생성되게끔

* fix: gitignore 추가, firebase config 수정

* feat: fcm 컴포넌트 수정
