+++ .gitignore
@@ -44,3 +44,4 @@ next-env.d.ts
 
 public/sw.js
 public/workbox-*.js
+public/firebase-config.js
