+++ scripts/generate-firebase-config.js
@@ -1,0 +1,19 @@
+const fs = require("fs");
+const path = require("path");
+require("dotenv").config();
+
+const config = {
+  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
+  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
+  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
+};
+
+const content = `self.FIREBASE_CONFIG = ${JSON.stringify(config, null, 2)};\n`;
+
+fs.writeFileSync(path.join(__dirname, "../public/firebase-config.js"), content);
+
+console.log("✅ firebase-config.js has been generated in /public");
